+++ services/sso/src/__tests__/dashboard/users/getMe.test.ts
@@ -2,20 +2,18 @@ import supertest from 'supertest'
 import { expect, it, describe, beforeAll, afterAll, afterEach } from 'vitest'
 import { pathRoot } from '../../../routes/routes'
 import { server, testUserData } from '../../globalSetup'
-import { client } from '../../../models/db'
 import { userSchema } from '../../../schemas'
 import { UserStatus } from '../../../schemas/users/userSchema'
+import { client } from '../../../db/client'
+import { dashboardLoginAndGetToken } from '../../helpers/testHelpers'
 
 const route = `${pathRoot.v1.dashboard.users}/me`
 let adminAuthToken = ''
 beforeAll(async () => {
-  const response = await supertest(server)
-    .post(`${pathRoot.v1.dashboard.auth}/login`)
-    .send({
-      dni: testUserData.admin.dni,
-      password: testUserData.admin.password,
-    })
-  ;[adminAuthToken] = response.header['set-cookie'][0].split(';')
+  adminAuthToken = await dashboardLoginAndGetToken(
+    testUserData.admin.dni,
+    testUserData.admin.password
+  )
 })
 
 afterAll(async () => {
@@ -27,13 +25,7 @@ afterAll(async () => {
   ])
 })
 
-describe('Testing get user endpoint', () => {
-  afterEach(async () => {
-    await client.query('UPDATE "user" SET status = $1 WHERE dni = $2', [
-      UserStatus.ACTIVE,
-      testUserData.admin.dni,
-    ])
-  })
+describe('Testing get users/me endpoint', () => {
   it('should succeed with a valid token', async () => {
     const response = await supertest(server)
       .get(route)
@@ -46,19 +38,13 @@ describe('Testing get user endpoint', () => {
         .safeParse(response.body).success
     ).toBeTruthy()
   })
-  it('should fail with 401 when cookies are not provided', async () => {
-    const response = await supertest(server).get(route)
-    expect(response.status).toBe(401)
-    expect(response.body.message).toBe('Invalid Credentials')
-  })
-  it('should fail with 401 error for invalid token', async () => {
-    adminAuthToken = 'invalid_token'
-    const response = await supertest(server)
-      .get(route)
-      .set('Cookie', [adminAuthToken])
-
-    expect(response.status).toBe(401)
-    expect(response.body.message).toBe('Invalid Credentials')
+})
+describe('User Status Handling in Get users/me Endpoint', () => {
+  afterEach(async () => {
+    await client.query('UPDATE "user" SET status = $1 WHERE dni = $2', [
+      UserStatus.ACTIVE,
+      testUserData.admin.dni,
+    ])
   })
   it('should fail when the logged-in admin loses "active" status', async () => {
     const agent = supertest.agent(server)