+++ services/sso/src/controllers/itineraries/list.ts
@@ -1,5 +1,5 @@
 import { Context, Middleware } from 'koa'
-import { client } from '../../models/db'
+import { client } from '../../db/client'
 
 export const listItineraries: Middleware = async (ctx: Context) => {
   const { rows: data } = await client.query('SELECT * FROM "itinerary";')